requirements.txt
Django==5.1.3
gunicorn
psycopg2-binary
whitenoise
dj-database-url
python-decouple
python-dotenv
requests
pillow
django-bootstrap-icons
django-ckeditor
django-allauth
django-simple-history
django-widget-tweaks
django-js-asset
google-generativeai
google-api-python-client
google-auth
opencv-python-headless
numpy

